[ActivityModeration] Handling of the delete form in NoticeOptionItems.
The majority of the changes concerns with making the plugin control
both the UI and logic aspects of the delete form in NoticeOptionItems.
Other changes are minor fixes or updates to the code visited during
the work.

ActivityModeration isn't still converted to a full "ActivityVerbHandlerPlugin",
that will require some time to properly do due to the changes done
to the core delete logic when this plugin was first introduced.

+++ plugins/ActivityModeration/ActivityModerationPlugin.php
@@ -18,7 +18,7 @@ class ActivityModerationPlugin extends ActivityVerbHandlerPlugin
 
     public function verbs()
     {
-        return array(ActivityVerb::DELETE);
+        return [ActivityVerb::DELETE];
     }
 
     public function onBeforePluginCheckSchema()
@@ -60,8 +60,7 @@ class ActivityModerationPlugin extends ActivityVerbHandlerPlugin
 
     protected function getActionTitle(ManagedAction $action, $verb, Notice $target, Profile $scoped)
     {
-        // FIXME: switch based on action type
-        return _m('TITLE', 'Notice moderation');
+        return _m('TITLE', 'Notice deleted');
     }
 
     protected function doActionPreparation(ManagedAction $action, $verb, Notice $target, Profile $scoped)
@@ -72,13 +71,17 @@ class ActivityModerationPlugin extends ActivityVerbHandlerPlugin
     protected function doActionPost(ManagedAction $action, $verb, Notice $target, Profile $scoped)
     {
         switch (true) {
-        case ActivityUtils::compareVerbs($verb, array(ActivityVerb::DELETE)):
-            // do whatever preparation is necessary to delete a verb
-            $target->deleteAs($scoped);
+        case ActivityUtils::compareVerbs($verb, $this->verbs()):
+            if (Event::handle('StartDeleteOwnNotice', [$scoped->getUser(), $target])) {
+                $target->deleteAs($scoped);
+                Event::handle('EndDeleteOwnNotice', [$scoped->getUser(), $target]);
+            }
             break;
         default:
             throw new ServerException('ActivityVerb POST not handled by plugin that was supposed to do it.');
         }
+
+        return false;
     }
 
     public function deleteRelated(Notice $notice)
@@ -205,10 +208,23 @@ class ActivityModerationPlugin extends ActivityVerbHandlerPlugin
     }
 
     protected function getActivityForm(ManagedAction $action, $verb, Notice $target, Profile $scoped) 
-    { 
-        if (!$scoped instanceof Profile || !($scoped->sameAs($target->getProfile()) || $scoped->hasRight(Right::DELETEOTHERSNOTICE))) {
-            throw new AuthorizationException(_('You are not allowed to delete other user\'s notices'));
+    {
+        return new InvisibleForm($action, $target);
+    }
+
+    public function onEndShowNoticeOptionItems(NoticeListItem $item) {
+        if (Event::handle('StartShowDeleteForm', [$item])) {
+            $profile = Profile::current();
+            $notice = empty($item->repeat) ? $item->notice : $item->repeat;
+
+            if ($profile !== null &&
+                !$notice->isVerb([ActivityVerb::DELETE]) &&
+                ($notice->profile_id == $profile->id || $profile->hasRight(Right::DELETEOTHERSNOTICE))) {
+                $form = new DeletenoticeForm($item->out, $item->notice);
+                $form->show();
+            }
+
+            Event::handle('EndShowDeleteForm', [$item]);
         }
-        return DeletenoticeForm($action, array('notice'=>$target));
-    } 
+    }
 }
